requirements.txt
discord.py>=2.3
requests
python-dotenv
tzdata
